requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
aiohttp==3.8.5
aiosignal==1.3.1
alt-profanity-check==1.3.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
cffi==1.15.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
datasets==2.14.4
dill==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
evaluate==0.4.0
fastapi==0.103.1
filelock==3.12.2
Flask==2.3.2
Flask-RESTful==0.3.10
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
Mako==1.2.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
numpy==1.24.4
orjson==3.9.5
packaging==23.1
pandas==2.0.3
passlib==1.7.4
pillow==10.0.0
psycopg==3.1.10
psycopg2==2.9.7
pyarrow==12.0.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
responses==0.18.0
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setfit==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
typing-extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
werkzeug==2.3.7
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
